feat:최근 검색어 및 인기 검색어 조회 인터페이스 추가
