+++ src/main/java/com/avengers/musinsa/domain/search/controller/PopularKeywordController.java
@@ -1,20 +1,17 @@
 package com.avengers.musinsa.domain.search.controller;
 
 import com.avengers.musinsa.domain.search.dto.PopularKeywordResponseDTO;
-import com.avengers.musinsa.domain.search.dto.ResultDTO;
-import com.avengers.musinsa.domain.search.service.PopularKeywordService;
+import com.avengers.musinsa.domain.search.service.PopularKeywordServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
-
 @RestController
 @RequestMapping("/api/v1/search")
 @RequiredArgsConstructor
 public class PopularKeywordController {
-    private final PopularKeywordService popularKeywordService;
+    private final PopularKeywordServiceImpl popularKeywordService;
 
     @GetMapping("/popular-keyword")
    public PopularKeywordResponseDTO findPopularKeyword(){
